+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 ### 新增
 - save last checkpoint when finished, interupted
 - use compatible lora statedict filename and format with huggingface, that can use diffusers pipeline to load the lora`# pipe.load_lora_weights(lora_weight, prefix="", adapter_name="lora_edit")`
+- upload lora weights to huggingface model
 
 ### 改进
 - 待添加...